Small changes to fix sketch upload issues
Even the rebuild of the original bootloader code has improved the situation and probably fixed the issues with "bricked" MINI. We have removed dead code - toggling with status LED. And added configuration of the higher byte of the baudrate register. Original bootloader assumes that the 0x00 value is set by the MCU reset, but it seems that in some cases it is not true.